SF Restaurant Analysis
San Fransisco Restaurants are inspected by the Health Department and
has developed an inspection report and scoring system. After conducting
an inspection of the facility, the Health Inspector calculates a score
based on the violations observed. Violations can fall into: 1.high risk
category
2.moderate risk category 3.low risk category
Violations that are directly related to the transmission of food borne
illnesses, the adulteration of food products and the contamination of
food-contact surfaces fall into high risk category.
Records specific violations that are of a moderate risk to the public
health and safety fall into moderate risk category.
Record violations that are low risk or have no immediate risk to the
public health and safety fall into low risk category.
The score card that will be issued by the inspector is maintained at
the food establishment and is available to the public in this dataset.

Here I have used Apache Spark on Databricks community Edition server
and Scala as the programming language.The text file in the folder
contains a link to the published notebook.